+++ azure-functions-powershell-library.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.2.32602.215
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AzureFunctionsSDK", "AzureFunctionsSDK\AzureFunctionsSDK.csproj", "{AA8C50CB-E5A8-4BF2-9EA7-EE57C8366916}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "azure-functions-powershell-library", "azure-functions-powershell-library\azure-functions-powershell-library.csproj", "{AA8C50CB-E5A8-4BF2-9EA7-EE57C8366916}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
