+++ azure-functions-powershell-library/AzureFunctions.PowerShell.SDK.ps1
@@ -1,3 +1,8 @@
+#	
+# Copyright (c) Microsoft. All rights reserved.	
+# Licensed under the MIT license. See LICENSE file in the project root for full license information.	
+#
+
 class Function : Attribute {
     [string]$name
