+++ azure-functions-powershell-library/BundledBindings/HttpTriggerBinding.cs
@@ -1,7 +1,12 @@
-﻿using Microsoft.Azure.Functions.PowerShellWorker;
+﻿//
+// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+//
+
+using Microsoft.Azure.Functions.PowerShellWorker;
 using System.Management.Automation.Language;
 
-namespace AzureFunctionsSDK.BundledBindings
+namespace AzureFunctions.PowerShell.SDK.BundledBindings
 {
     public class HttpTriggerBinding : IInputBinding
     {
