+++ azure-functions-powershell-library/BundledBindings/IBinding.cs
@@ -1,7 +1,12 @@
-﻿using Microsoft.CodeAnalysis.CSharp.Syntax;
+﻿//
+// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+//
+
+using Microsoft.CodeAnalysis.CSharp.Syntax;
 using System.Management.Automation.Language;
 
-namespace AzureFunctionsSDK.BundledBindings
+namespace AzureFunctions.PowerShell.SDK.BundledBindings
 {
     public abstract class IBinding
     {
