+++ azure-functions-powershell-library/BundledBindings/IOutputBinding.cs
@@ -1,11 +1,16 @@
-﻿using System;
+﻿//
+// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+//
+
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Management.Automation.Language;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace AzureFunctionsSDK.BundledBindings
+namespace AzureFunctions.PowerShell.SDK.BundledBindings
 {
     public abstract class IOutputBinding : IBinding
     {
