+++ azure-functions-powershell-library/BundledBindings/OrchestrationTriggerBinding.cs
@@ -1,13 +1,18 @@
-﻿using Microsoft.Azure.Functions.PowerShellWorker;
+﻿//
+// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+//
+
+using Microsoft.Azure.Functions.PowerShellWorker;
 using System.Management.Automation.Language;
 
-namespace AzureFunctionsSDK.BundledBindings
+namespace AzureFunctions.PowerShell.SDK.BundledBindings
 {
-    public class EventGridTriggerBinding : IInputBinding
+    public class OrchestrationTriggerBinding : IInputBinding
     {
-        public override string BindingAttributeName => "EventGridTrigger";
+        public override string BindingAttributeName => "OrchestrationTrigger";
 
-        public override string BindingType => "eventGridTrigger";
+        public override string BindingType => "orchestrationTrigger";
 
         public override BindingInformation ExtractBinding(AttributeAst attribute, ParameterAst parameter)
         {
