+++ azure-functions-powershell-library/GetFunctionsMetadataCommand.cs
@@ -1,5 +1,9 @@
-﻿using AzureFunctionsSDK;
-using AzureFunctionsSDK.BundledBindings;
+﻿//
+// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+//
+
+using AzureFunctions.PowerShell.SDK;
 using Microsoft.Azure.Functions.PowerShellWorker;
 using System.Management.Automation;
