+++ azure-functions-powershell-library/azure-functions-powershell-library.csproj
